add context

+++ .pipelines/restore.sh
@@ -18,24 +18,64 @@ echo -n $BUILD_BUILDNUMBER | tee ./EV2Specs/BuildVer.txt
 
 
 #new start
-ARCHS=("Amd64" "Arm64")
+# ARCHS=("Amd64" "Arm64")
+# for arch in "${ARCHS[@]}"; do
+#     mkdir $arch
+#     IMAGE_NAMES=("agentInit" "agent" "operator")
+#     # IMAGE_NAMES=("agent")
+#     #这个文件夹的名字应该取决于你如何name的pipeline，应该不是UI上的是yaml里的
+#     for image in "${IMAGE_NAMES[@]}"; do
+#         ORIGINAL_DIRECTORY="../../../../retina-oss-build/drop_build_${image}Linux${arch}ImageBuild"
+
+#         for file in "$ORIGINAL_DIRECTORY"/*; do
+#             if [[ "$file" == *.tar.gz ]]; then
+#                 gunzip "$file"  # 解压缩
+#                 file="${file%.gz}"  # 去掉 .gz 后缀
+#                 mv "$file" "./$arch/"  # 直接移动到新目录
+#             fi
+#         done
+#         echo "$arch Folder Contents"
+#         ls -alF $arch
+#     done
+# done
+echo "Listing contents of ../../../../"
+ls ../../../../
+
 for arch in "${ARCHS[@]}"; do
-    mkdir $arch
+    mkdir -p "$arch"
     IMAGE_NAMES=("agentInit" "agent" "operator")
-    # IMAGE_NAMES=("agent")
-    #这个文件夹的名字应该取决于你如何name的pipeline，应该不是UI上的是yaml里的
+
     for image in "${IMAGE_NAMES[@]}"; do
         ORIGINAL_DIRECTORY="../../../../retina-oss-build/drop_build_${image}Linux${arch}ImageBuild"
+        
+        # 检查目录是否存在
+        if [ ! -d "$ORIGINAL_DIRECTORY" ]; then
+            echo "Error: Directory does not exist - $ORIGINAL_DIRECTORY"
+            continue
+        fi
+        
+        echo "Processing directory: $ORIGINAL_DIRECTORY"
+        ls -alF "$ORIGINAL_DIRECTORY"
 
         for file in "$ORIGINAL_DIRECTORY"/*; do
+            echo "Processing file: $file"
             if [[ "$file" == *.tar.gz ]]; then
-                gunzip "$file"  # 解压缩
-                file="${file%.gz}"  # 去掉 .gz 后缀
-                mv "$file" "./$arch/"  # 直接移动到新目录
+                echo "Decompressing file: $file"
+                gunzip "$file" && echo "Decompressed: $file"
+                file="${file%.gz}"
+                if [ -f "$file" ]; then
+                    echo "Moving file: $file to ./$arch/"
+                    mv "$file" "./$arch/"
+                else
+                    echo "Error: File $file not found after decompression."
+                fi
+            else
+                echo "Skipping non-matching file: $file"
             fi
         done
-        echo "$arch Folder Contents"
-        ls -alF $arch
+
+        echo "$arch Folder Contents:"
+        ls -alF "$arch"
     done
 done
